refactor: 회고 생성, 템플릿 폴더구조 변경
