+++ apps/web/src/app/desktop/component/retrospect/template/recommend/Done.tsx
@@ -12,8 +12,8 @@ import { chooseParticle } from "@/utils/retrospect/chooseParticle";
 import { useFunnelModal } from "@/hooks/useFunnelModal";
 import { TemplateCard } from "../card/TemplateCard";
 import { useActionModal } from "@/hooks/useActionModal";
-import ChoiceTemplate from "..";
-import { RetrospectCreate } from "@/app/desktop/retrospectCreate/RetrospectCreate";
+import { ChoiceTemplate } from "@/app/desktop/component/retrospect/choice";
+import { RetrospectCreate } from "@/app/desktop/component/retrospectCreate";
 
 function RecommendDone() {
   const setRetrospectValue = useSetAtom(retrospectInitialState);
