+++ apps/web/src/app/desktop/component/retrospectCreate/index.tsx
@@ -1,15 +1,15 @@
 import { createContext, useCallback } from "react";
-import { ConfirmDefaultTemplate } from "../component/retrospectCreate/steps/ConfirmDefaultTemplate";
 import { ProgressBar } from "@/component/common/ProgressBar";
 import { css } from "@emotion/react";
-import { DueDate } from "../component/retrospectCreate/steps/DueDate";
 import { useMultiStepForm } from "@/hooks/useMultiStepForm";
 import { usePostRetrospectCreate } from "@/hooks/api/retrospect/create/usePostRetrospectCreate";
 import { useAtomValue } from "jotai";
 import { retrospectCreateAtom } from "@/store/retrospect/retrospectCreate";
 import { REQUIRED_QUESTIONS } from "@/component/retrospectCreate/customTemplate/questions.const";
 import { retrospectInitialState } from "@/store/retrospect/retrospectInitial";
-import { MainInfo } from "../component/retrospectCreate/steps/MainInfo";
+import { MainInfo } from "./steps/MainInfo";
+import { DueDate } from "./steps/DueDate";
+import { ConfirmDefaultTemplate } from "./steps/ConfirmDefaultTemplate";
 
 const PAGE_STEPS = ["confirmTemplate", "mainInfo", "dueDate"] as const;
 const CUSTOM_TEMPLATE_STEPS = ["confirmDefaultTemplate", "editQuestions", "confirmEditTemplate"] as const;
