+++ apps/web/src/app/desktop/component/retrospectCreate/steps/MainInfo.tsx
@@ -1,4 +1,3 @@
-import { RetrospectCreateContext } from "@/app/desktop/retrospectCreate/RetrospectCreate";
 import { ButtonProvider } from "@/component/common/button";
 import { Header } from "@/component/common/header";
 import { Input, InputLabelContainer, Label, TextArea } from "@/component/common/input";
@@ -9,6 +8,7 @@ import { retrospectCreateAtom } from "@/store/retrospect/retrospectCreate";
 import { css } from "@emotion/react";
 import { useAtom } from "jotai";
 import { useContext } from "react";
+import { RetrospectCreateContext } from "..";
 
 export function MainInfo() {
   const { goNext, goPrev } = useContext(RetrospectCreateContext);
